Refactor file structure
