+++ src/components/admin/order/realtime/NotPaidOrderCard.tsx
@@ -3,13 +3,13 @@ import { Order } from '@@types/index';
 import AppLabel from '@components/common/label/AppLabel';
 import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
-import { RiCloseLargeLine, RiCheckLine, RiArrowRightSLine } from '@remixicon/react';
+import { RiArrowRightSLine, RiCheckLine, RiCloseLargeLine } from '@remixicon/react';
 import { expandButtonStyle } from '@styles/buttonStyles';
 import { colFlex, rowFlex } from '@styles/flexStyles';
 import OrderSummaryContents from './OrderSummaryContents';
 import useAdminOrder from '@hooks/admin/useAdminOrder';
 import { useParams } from 'react-router-dom';
-import OrderDetailModal from '@components/admin/order/modal/OrderDetailModal';
+import OrderDetailModal from '@components/admin/order/realtime/modal/OrderDetailModal';
 import { areOrdersEquivalent } from '@utils/MemoCompareFunction';
 import useDelayTime from '@hooks/useDelayTime';
 import useModal from '@hooks/useModal';
