+++ src/components/admin/order/realtime/ServedOrderCard.tsx
@@ -3,17 +3,16 @@ import { Order } from '@@types/index';
 import AppLabel from '@components/common/label/AppLabel';
 import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
-import { RiArrowRightSLine, RiCheckLine } from '@remixicon/react';
 import { expandButtonStyle } from '@styles/buttonStyles';
 import { colFlex, rowFlex } from '@styles/flexStyles';
+import OrderSummaryContents from './OrderSummaryContents';
 import useAdminOrder from '@hooks/admin/useAdminOrder';
 import { useParams } from 'react-router-dom';
 import RollBackSvg from '@resources/svg/RollBackSvg';
-import OrderDetailModal from '@components/admin/order/modal/OrderDetailModal';
-import OrderItemList from '@components/admin/order/OrderItemList';
+import OrderDetailModal from '@components/admin/order/realtime/modal/OrderDetailModal';
 import { areOrdersEquivalent } from '@utils/MemoCompareFunction';
-import useDelayTime from '@hooks/useDelayTime';
 import useModal from '@hooks/useModal';
+import { RiArrowRightSLine } from '@remixicon/react';
 
 const CardContainer = styled.div`
   ${colFlex({ justify: 'center', align: 'center' })}
@@ -36,7 +35,7 @@ const CardContainer = styled.div`
 const CardContents = styled.div`
   ${colFlex({ justify: 'space-between', align: 'center' })}
   width: 100%;
-  height: 90%;
+  height: 85%;
 `;
 
 const OrderInfoContainer = styled.div`
@@ -48,7 +47,6 @@ const OrderInfoContainer = styled.div`
   width: 100%;
   cursor: pointer;
 `;
-
 const HeaderContainer = styled.div`
   ${rowFlex({ justify: 'space-between', align: 'start' })}
   width: 100%;
@@ -65,7 +63,7 @@ const CheckButtonContainer = styled.div`
   width: 55%;
 `;
 
-const CheckIcon = styled(RiCheckLine)`
+const RightIcon = styled(RiArrowRightSLine)`
   width: 25px;
   height: 25px;
   ${expandButtonStyle}
@@ -77,12 +75,6 @@ const RollBackIcon = styled(RollBackSvg)`
   ${expandButtonStyle}
 `;
 
-const RightIcon = styled(RiArrowRightSLine)`
-  width: 25px;
-  height: 25px;
-  ${expandButtonStyle}
-`;
-
 interface OrderCardProps {
   order: Order;
 }
@@ -91,18 +83,13 @@ const arePropsEqual = (prevProps: OrderCardProps, nextProps: OrderCardProps) => 
   return areOrdersEquivalent(prevProps.order, nextProps.order);
 };
 
-function PaidOrderCard({ order }: OrderCardProps) {
+function ServedOrderCard({ order }: OrderCardProps) {
   const { workspaceId } = useParams<{ workspaceId: string }>();
-  const { serveOrder, refundOrder } = useAdminOrder(workspaceId);
-  const { delayMinutes } = useDelayTime({ date: order.createdAt });
+  const { payOrder } = useAdminOrder(workspaceId);
   const { isModalOpen, openModal, closeModal } = useModal();
 
-  const checkClickHandler = () => {
-    serveOrder(order.id);
-  };
-
   const rollBackClickHandler = () => {
-    refundOrder(order.id);
+    payOrder(order.id);
   };
 
   const orderInfoClickHandler = () => {
@@ -119,15 +106,16 @@ function PaidOrderCard({ order }: OrderCardProps) {
                 <AppLabel color={Color.BLACK} size={17} style={{ fontWeight: 800 }}>
                   {`테이블 ${order.tableNumber}`}
                 </AppLabel>
-                <AppLabel color={Color.BLACK} size={13}>{`${delayMinutes}분 전 주문`}</AppLabel>
+                <AppLabel color={Color.BLACK} size={13}>
+                  {`주문 번호 ${order.orderNumber}`}
+                </AppLabel>
               </TitleContainer>
               <RightIcon onClick={openModal} />
             </HeaderContainer>
-            <OrderItemList order={order} />
+            <OrderSummaryContents contents={order} />
           </OrderInfoContainer>
           <OrderDetailModal order={order} isModalOpen={isModalOpen} closeModal={closeModal} />
           <CheckButtonContainer>
-            <CheckIcon onClick={checkClickHandler} />
             <RollBackIcon onClick={rollBackClickHandler} />
           </CheckButtonContainer>
         </CardContents>
@@ -136,4 +124,4 @@ function PaidOrderCard({ order }: OrderCardProps) {
   );
 }
 
-export default memo(PaidOrderCard, arePropsEqual);
+export default memo(ServedOrderCard, arePropsEqual);
