+++ src/components/admin/order/realtime/modal/OrderDetailModal.tsx
@@ -2,9 +2,9 @@ import { Order } from '@@types/index';
 import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
 import { colFlex } from '@styles/flexStyles';
-import OrderModalFooterContents from '@components/admin/order/modal/OrderModalFooterContents';
-import OrderModalMainContents from '@components/admin/order/modal/OrderModalMainContents';
-import OrderModalHeaderContents from '@components/admin/order/modal/OrderModalHeaderContents';
+import OrderModalFooterContents from '@components/admin/order/realtime/modal/OrderModalFooterContents';
+import OrderModalMainContents from '@components/admin/order/realtime/modal/OrderModalMainContents';
+import OrderModalHeaderContents from '@components/admin/order/realtime/modal/OrderModalHeaderContents';
 import { createPortal } from 'react-dom';
 import { useEffect } from 'react';
 import useModal from '@hooks/useModal';
