+++ src/pages/admin/order/AdminOrderStatistics.tsx
@@ -9,11 +9,11 @@ import AppContainer from '@components/common/container/AppContainer';
 import { colFlex, rowFlex } from '@styles/flexStyles';
 import { OrderStatus } from '@@types/index';
 import { tabletMediaQuery } from '@styles/globalStyles';
-import OrderStatisticsNavBarChildren from '@components/admin/order/OrderStatisticsNavBarChildren';
+import OrderStatisticsNavBarChildren from '@components/admin/order/statistic/OrderStatisticsNavBarChildren';
 import { Color } from '@resources/colors';
-import TotalOrder from '@components/admin/order/TotalOrder';
-import ProductStatistics from '@components/admin/order/ProductStatistics';
-import OrderPriceStatistics from '@components/admin/order/OrderPriceStatistics';
+import TotalOrder from '@components/admin/order/statistic/TotalOrder';
+import ProductStatistics from '@components/admin/order/statistic/ProductStatistics';
+import OrderPriceStatistics from '@components/admin/order/statistic/OrderPriceStatistics';
 import { lineSeedKrFont } from '@styles/fonts';
 import dayjs from 'dayjs';
 import { dateConverter } from '@utils/FormatDate';
